requirements.txt
flask
jinja2
uvicorn
quart
htmlmin2
csscompressor
jsmin
requests